pyproject.toml
[project]
name = "opsinsight"
version = "0.1.0"
dependencies = [
    "google-genai"
]

[tool.setuptools]
py-modules = ["ops", "utils"]

[project.scripts]
ops = "ops:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"